Update hw6.md
(Removed the following because it was not in out original requirements) : |11 | Revised :pencil2: and Completed :white_check_mark: | Hompage | Statistics Dashboards | Next to the displayed static data, display a bar chart of new cases per day for the past 7 days. | The API is unable to provide daily case data for the past 7 days, so the bar chart shows total cases as of two weeks ago, one week ago, and one day ago. |